backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
pydantic-settings
python-dotenv
openai
langchain
langchain-openai
langchain-community
chromadb
python-multipart
pypdf
docx2txt
requests
imageio-ffmpeg
PyMuPDF
python-docx
python-pptx
youtube-transcript-api